+++ src/lib/opengpex/core/advanced/commands/frame/revert.ts
@@ -21,7 +21,8 @@
 
 import { EditorCommand, EditorContextValue, asLocalShape, Layer } from '@opengpex/editor/core/types';
 import { LayerFactory } from '@opengpex/editor/core/layer';
-import { VIEWPORT_FIT_PADDING } from '@opengpex/editor/core/helpers/presets';
+import { presets } from '@opengpex/editor/core/helpers/preferences';
+const VIEWPORT_FIT_PADDING = presets.get('VIEWPORT_FIT_PADDING');
 import * as P from '@opengpex/editor/core/advanced/protocols';
 
 export const FrameRevertCommand = {